AirconMoa-Android-6 CreateUserActivity-소셜로그인 및 파이어베이스-관련 설정

+++ app/src/main/java/com/example/airconmoa/ui/join_user/EnterNicknameActivity.kt
@@ -1,20 +1,26 @@
 package com.example.airconmoa.ui.join_user
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.view.View
 import android.widget.Toast
 import com.example.airconmoa.config.BaseActivityVB
-import com.example.airconmoa_android.R
-import com.example.airconmoa_android.databinding.ActivityEnterEmailBinding
-import com.example.airconmoa_android.databinding.ActivityEnterNicknameBinding
+import com.example.airconmoa.R
+import com.example.airconmoa.databinding.ActivityEnterNicknameBinding
+import com.example.airconmoa.util.FirebaseAuthUtils
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
 
 class EnterNicknameActivity : BaseActivityVB<ActivityEnterNicknameBinding>(ActivityEnterNicknameBinding::inflate) {
 
+    private lateinit var auth: FirebaseAuth
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        auth = Firebase.auth
 
         binding.enterNicknameBackIv.setOnClickListener {
             finish()
@@ -35,23 +41,46 @@ class EnterNicknameActivity : BaseActivityVB<ActivityEnterNicknameBinding>(Activ
                 showCustomToast("닉네임은 1~10글자 이내로 입력해주세요")
             }
             else {
-                val intent = Intent(this, FinishJoinActivity::class.java)
-                intent.putExtra("nickname", binding.enterNicknameEt.text.toString())
                 // 유저의 정보
                 val sharedPreferences = getSharedPreferences("userInfo", MODE_PRIVATE)
                 val userEmail = sharedPreferences.getString("userEmail", null)
                 val userPassword = sharedPreferences.getString("userPassword", null)
                 val userPhoneNumber = sharedPreferences.getString("userPhoneNumber", null)
-                Log.d("userInfo", userEmail!! + "/" +userPassword!! + "/"
-                        + userPhoneNumber!! + "/" + binding.enterNicknameEt.text.toString())
+                val userNickname = binding.enterNicknameEt.text.toString()
 
-                // sharedPreferences내에 저장된 유저의 모든 데이터를 삭제함
-                val editor = sharedPreferences.edit()
-                editor.clear()
-                editor.apply()
-                binding.enterNicknameErrorTv.visibility = View.INVISIBLE
-                startActivity(intent)
-                overridePendingTransition(R.anim.slide_right_enter, R.anim.slide_right_exit)
+                if(userEmail != null && userPassword != null && userPhoneNumber != null) {
+                    Log.d("userInfo", userEmail!! + "/" + userPassword!! + "/"
+                            + userPhoneNumber!! + "/" + userNickname)
+
+                    auth.createUserWithEmailAndPassword(userEmail, userPassword)
+                        .addOnCompleteListener(this) { task ->
+                            if (task.isSuccessful) {
+                                showCustomToast("가입을 환영합니다!")
+                                // sharedPreferences내에 저장된 유저의 모든 데이터를 삭제함
+                                val editor = sharedPreferences.edit()
+                                editor.clear()
+                                editor.apply()
+                                binding.enterNicknameErrorTv.visibility = View.INVISIBLE
+                                val intent = Intent(this, FinishJoinActivity::class.java)
+                                intent.putExtra("nickname", userNickname)
+                                startActivity(intent)
+                                overridePendingTransition(R.anim.slide_right_enter, R.anim.slide_right_exit)
+                            } else {
+                                Log.d("EnterNicknameActivity", "회원가입 실패")
+                            }
+                        }
+                }
+                else {
+                    if(FirebaseAuthUtils.getUid() != null) {
+                        val intent = Intent(this, FinishJoinActivity::class.java)
+                        intent.putExtra("nickname", userNickname)
+                        startActivity(intent)
+                        overridePendingTransition(R.anim.slide_right_enter, R.anim.slide_right_exit)
+                    }
+                    else {
+                        showCustomToast("회원가입 정보가 올바른지 다시 확인해주세요")
+                    }
+                }
             }
         }
         setFullScreen()
